% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectureslides.R
\name{lectureslides}
\alias{lectureslides}
\title{Beamer Lecture Slides}
\usage{
lectureslides(lang = "en", towers = T, template = "default",
  colorlinks = T, tables = T, institute = T, blockstyle = "box",
  shadecolor = "default", toc = TRUE, slide_level = 2,
  incremental = FALSE, fig_width = 10, fig_height = 7,
  fig_crop = TRUE, fig_caption = TRUE, dev = "pdf",
  df_print = "default", theme = "default", colortheme = "default",
  fonttheme = "default", highlight = "default", keep_tex = FALSE,
  latex_engine = "pdflatex", citation_package = c("none", "natbib",
  "biblatex"), self_contained = TRUE, includes = NULL,
  md_extensions = "+fancy_lists", pandoc_args = NULL)
}
\arguments{
\item{lang}{\code{character}. Language of the presentation. One of \code{"de"} or \code{"en"}. Defaults to \code{"en"}.}

\item{towers}{\code{logical}. Should the title page include a gradient logo of the university buildings.}

\item{template}{\code{character}. The template to be used. Defaults to \code{default}}

\item{institute}{\code{logical} or \code{character}. Should the name of the institution be included on the title page.
Defaults to \code{T} and includes the name of the University Duisburg-Essen in the language specified by \code{lang}.
If a \code{character} object is provided, this string is shown underneath the title.}

\item{blockstyle}{\code{character}. The style of the beamer block environments (theorems, examples, etc.). One of \code{"box"} or \code{"blank"}.
Defaults to \code{"box"}.}

\item{shadecolor}{\code{numeric}. The shade of grey used as the background color for code chunks. A number between 0 and 1.}

\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Beamer Lecture Slides
}
\details{
Include a logo using the YAML option \code{logo}.
}
\examples{
\dontrun{ library(rmarkdown) draft("MyLecture.Rmd", template = "lectureslides",
package = "rmemo") }
}
