% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s}
\alias{s}
\title{Function to auto print inline code including the numerical results}
\usage{
s(x)
}
\arguments{
\item{x}{A R expression.}
}
\description{
Function to auto print inline code including the numerical results
}
\examples{
s(pi)
}
