% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectureslides.R
\name{lectureslides}
\alias{lectureslides}
\title{Beamer Lecture Slides}
\usage{
lectureslides(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}

Format for creating basic memos.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Beamer Lecture Slides
}
\details{
Include a logo using the YAML option \code{logo}.
}
\examples{
\dontrun{ library(rmarkdown) draft("MyLecture.Rmd", template = "lectureslides",
package = "rmemo") }
}
