% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectureslides.R
\name{lectureslides}
\alias{lectureslides}
\title{Beamer Lecture Slides}
\usage{
lectureslides(lang = "en", towers = T, template = "default",
  colorlinks = T, tables = T, institute = T, blockstyle = "box",
  toc = TRUE, slide_level = 2, incremental = FALSE, fig_width = 10,
  fig_height = 7, fig_crop = TRUE, fig_caption = TRUE, dev = "pdf",
  df_print = "default", theme = "default", colortheme = "default",
  fonttheme = "default", highlight = "default", keep_tex = FALSE,
  latex_engine = "pdflatex", citation_package = c("none", "natbib",
  "biblatex"), self_contained = TRUE, includes = NULL,
  md_extensions = NULL, pandoc_args = NULL)
}
\arguments{
\item{template}{Pandoc template to use for rendering. Pass "default" to use
the rmarkdown package default template; pass \code{NULL} to use pandoc's
built-in template; pass a path to use a custom template that you've
created.  See the documentation on
\href{{http://pandoc.org/README.html}{pandoc online documentation}}
for details on creating custom templates.}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{dev}{Graphics device to use for figure output (defaults to pdf)}

\item{df_print}{Method to be used for printing data frames. Valid values
include "default", "kable", "tibble", and "paged". The "default" method uses
\code{print.data.frame}. The "kable" method uses the
\code{\link[knitr:kable]{knitr::kable}} function. The "tibble" method uses
the \pkg{tibble} package to print a summary of the data frame. The "paged"
method creates a paginated HTML table (note that this method is only valid
for formats that produce HTML). In addition
to the named methods you can also pass an arbitrary function to be used
for printing data frames. You can disable the df_print behavior entirely
by setting the option \code{rmarkdown.df_print} to \code{FALSE}.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{Arguments to \code{rmarkdown::pdf_document}

Format for creating basic memos.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Beamer Lecture Slides
}
\details{
Include a logo using the YAML option \code{logo}.
}
\examples{
\dontrun{ library(rmarkdown) draft("MyLecture.Rmd", template = "lectureslides",
package = "rmemo") }
}
