% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectureslides.R
\name{lectureslides}
\alias{lectureslides}
\title{Beamer Lecture Slides}
\usage{
lectureslides(
  lang = "en",
  footerstyle = c("default", "colored"),
  towers = T,
  template = "default",
  colorlinks = T,
  tables = T,
  institute = T,
  blockstyle = c("box", "blank"),
  shadecolor = "default",
  logo = "default",
  quiz_logo = "default",
  slide_level = 2,
  incremental = FALSE,
  fig_width = 10,
  fig_height = 7,
  fig_crop = TRUE,
  fig_caption = TRUE,
  dev = "tikz",
  df_print = "default",
  theme = "default",
  colortheme = "default",
  fonttheme = "default",
  highlight = "default",
  keep_tex = TRUE,
  keep_md = TRUE,
  keep_aux = TRUE,
  latex_engine = "pdflatex",
  citation_package = c("none", "biblatex"),
  self_contained = TRUE,
  link_citations = TRUE,
  includes = NULL,
  md_extensions = "+fancy_lists",
  pandoc_args = NULL
)
}
\arguments{
\item{lang}{\code{character}. Language of the presentation. One of
\code{"de"} or \code{"en"}. Defaults to \code{"en"}.}

\item{footerstyle}{\code{character} One of "default" or "colored".}

\item{towers}{\code{logical}. Should the title page include a gradient logo
of the university buildings.}

\item{template}{\code{character}. The template to be used. Defaults to
\code{default}}

\item{colorlinks}{\code{logical} Should links be colored.}

\item{tables}{\code{logical} Should the latex packages "longtables" and "bookstabs" be included.}

\item{institute}{\code{logical} or \code{character}. Should the name of the
institution be included on the title page. Defaults to \code{T} and
includes the name of the University Duisburg-Essen in the language
specified by \code{lang}. If a \code{character} object is provided, this
string is shown underneath the title.}

\item{blockstyle}{\code{character}. The style of the beamer block
environments (theorems, examples, etc.). One of \code{"box"} or
\code{"blank"}. Defaults to \code{"box"}.}

\item{shadecolor}{\code{numeric}. The shade of grey used as the background
color for code chunks. A number between 0 and 1.}

\item{logo}{\code{character} Path to a logo for the slides. If not set otherwise, a UDE logo is used.}

\item{quiz_logo}{\code{character} Path to a logo for the slides. If not set otherwise, a quiz academy logo is used.}

\item{slide_level}{\code{numeric} Depth of slide numbering. Defaults to 2.}

\item{incremental}{\code{logical} Should the slides render incrementally. Defaults to \code{FALSE}.}

\item{fig_width}{\code{numeric} Default width of figures in inches. Defaults to 10.}

\item{fig_height}{\code{numeric} Default height of figures in inches. Defaults to 7.}

\item{fig_crop}{\code{logical} Should white space be cropped from figures.}

\item{fig_caption}{\code{logical} Should figure captions be used. Defaults to \code{TRUE}.}

\item{dev}{\code{character} The graphic device to be used. Defaults to "tikz".}

\item{df_print}{See \link[rmarkdown]{output_format}.}

\item{theme}{\code{character} The theme for the slides.}

\item{colortheme}{\code{character} The color theme for the slides.}

\item{fonttheme}{\code{character} The font theme for the slides.}

\item{highlight}{\code{character} One of \code{rmarkdown:::highlighters()}.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF
(applies only to 'latex' and 'beamer' target formats)}

\item{keep_md}{Keep the markdown file generated by knitting. Note that if
this is \code{TRUE} then \code{clean_supporting} will always be
\code{FALSE}.}

\item{latex_engine}{LaTeX engine to producing PDF output (applies only to
'latex' and 'beamer' target formats)}

\item{citation_package}{The citation package to be used. One of c("none", "natbib", "biblatex")}

\item{self_contained}{Produce stand-alone intermediate files.}

\item{includes}{Files that should be included via pandoc. See pandoc arguments.}

\item{md_extensions}{Markdown extensions for pandoc. Default to \code{+fancy_lists}.}

\item{pandoc_args}{Further arguments passed on to pandoc.}

\item{link-citations}{\code{logical} Should links be clickable. Defaults to \code{TRUE}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Beamer Lecture Slides
}
\section{Available Environment for Box}{

\tabular{ll}{
  \strong{Command} \tab \strong{Purpose} \cr
  \code{thmn}  \tab Theorem                    \cr
  \code{lem}   \tab Lemma                      \cr
  \code{cor}   \tab Corollary                  \cr
  \code{defn}  \tab Definition                 \cr
  \code{ass}   \tab Assumption (Annahme)       \cr
  \code{rem}   \tab Remark (Anmerkung)         \cr
  \code{prop}  \tab Proposition                \cr
  \code{xmpl}  \tab Example (Beispiel)         \cr
  \code{que}   \tab Question (Frage)           \cr
  \code{exe}   \tab Exercise (Aufgabe)         \cr
  \code{mthmn} \tab Mathematical Theorem block \cr
}
}

\section{Box‐color variables}{

\tabular{lll}{
  \strong{Variable}                \tab \strong{Use}                       \tab \strong{Default Color} \cr
  \code{theorem_box_color}         \tab Theorem                             \tab blue   \cr
  \code{lemma_box_color}           \tab Lemma                               \tab blue   \cr
  \code{corollary_box_color}       \tab Corollary                           \tab blue   \cr
  \code{definition_box_color}      \tab Definition                          \tab green  \cr
  \code{assumption_box_color}      \tab Assumption (Annahme)                \tab blue   \cr
  \code{remark_box_color}          \tab Remark (Anmerkung)                  \tab blue   \cr
  \code{proposition_box_color}     \tab Proposition                         \tab blue   \cr
  \code{example_box_color}        \tab Example (Beispiel)                  \tab yellow \cr
  \code{question_box_color}           \tab Question (Frage)                    \tab blue   \cr
  \code{exercise_box_color}        \tab Exercise (Aufgabe)                  \tab blue   \cr
  \code{math_theorem_box_color}    \tab Mathematical Theorem                \tab blue   \cr
}
}

\examples{

```{Color Box Variable Use}
title: "Induktive Statistik"
author: "Prof. Dr. Christoph Hanck"
date: "Sommersemester 2025"
output: 
  runidue::lectureslides:
  lang: "en" 
keep_tex: true
exercise_box_color: yellow
```

```{Environment Use}
\xmpl[Heading of the example]
  The example itself …
\endxmpl
```

```{Long Enironment Use}
# Example of a long environment
\xmpl[Long Environments]
A very long example which may not fit the slide
\endxmpl
# Continuation of the previous example
\xmpl[*]
This is the continuation of the long example in the next slide.
\endxmpl
```

```{Kahoot Box Use}
# Example of kahoot box
\kahoot {Kahoot Heading} Kahoot Contents \endkahoot
```

\dontrun{ library(rmarkdown) draft("MyLecture.Rmd", template = "lectureslides",
package = ("runidue") }
}
